+++ include/nxtgm/optimizers/gm/discrete/movemaker.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <nxtgm/functions/discrete_constraints.hpp>
+#include <nxtgm/functions/unique_labels_constraint_function.hpp>
 #include <nxtgm/models/gm/discrete_gm.hpp>
 #include <nxtgm/nxtgm.hpp>
 #include <nxtgm/optimizers/optimizer_base.hpp>
