+++ python/src/nxtgm/functions/export_discrete_constraint_functions.cpp
@@ -1,5 +1,6 @@
+#include <nxtgm/functions/array_constraint_function.hpp>
 #include <nxtgm/functions/discrete_constraint_function_base.hpp>
-#include <nxtgm/functions/discrete_constraints.hpp>
+#include <nxtgm/functions/unique_labels_constraint_function.hpp>
 #include <pybind11/pybind11.h>
 #include <xtensor-python/pyarray.hpp>
 #include <xtensor-python/pytensor.hpp>
